PopupPrompt Class Implementation
There is now a separate class for popup prompts making it much easier to create them. Pass in the necessary parameters and you can now give any SFElement the ability to give you options upon button press. This class also makes use of a @Binding variable to automatically hide the prompt when anywhere else on the screen is pressed.